print routes during start as well
